requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.2